pyproject.toml
[tool.poetry]
name = "haptools"
version = "0.0.1"
description = "Simulate phenotypes for GWAS and subsequent fine-mapping"
authors = ["Arya Massarat", "Michael Lamkin"]
license = "MIT"
repository = "https://github.com/aryarm/haptools"
homepage = "https://github.com/aryarm/haptools"

[tool.poetry.dependencies]
python = "^3.7"
click = "^8.0.3"
pysam = "^0.18.0"
cyvcf2 = "^0.30.14"

# docs
# these belong in dev-dependencies, but RTD doesn't support that yet -- see
# https://github.com/readthedocs/readthedocs.org/issues/4912
Sphinx = {version = "^4.3.2", optional = true }
sphinx-autodoc-typehints = {version = "^1.12.0", optional = true }
sphinx-rtd-theme = {version = "^1.0.0", optional = true }
numpydoc = {version = "^1.1.0", optional = true }
sphinx-click = {version = "^3.0.2", optional = true }
brewer2mpl = "^1.4.1"
pytabix = "^0.1"

[tool.poetry.dev-dependencies]

# docs
# these belong in dev-dependencies, but RTD doesn't support that yet -- see
# https://github.com/readthedocs/readthedocs.org/issues/4912
Sphinx = {version = "^4.3.2" }
sphinx-autodoc-typehints = {version = "^1.12.0" }
sphinx-rtd-theme = {version = "^1.0.0" }
numpydoc = {version = "^1.1.0" }
sphinx-click = {version = "^3.0.2" }
pytest = "^6.2.5"
black = "^21.12b0"
ipython = "^7.30.1"

[tool.poetry.extras]
docs = [
    "sphinx",
    "sphinx-autodoc-typehints",
    "sphinx-rtd-theme",
    "numpydoc",
    "sphinx-click"
]

test = [
    "pytest"
]

[tool.poetry.scripts]
haptools = 'haptools.__main__:main'

[tool.black]
line-length = 88
experimental-string-processing = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
